Add SQLite support
This adds SQLite support, but a few queries are broken because they're PostgreSQL-exclusive/specific.

I don't see enough value (at least for now) in making the queries harder to read (or the code around them) "just" to enable SQLite support, so I'll abandon this for now, but push the code up in case it helps anyone else add this kind of support (and then change the queries for themselves).

Related to #49